update page fetch to react query
