+++ package.json
@@ -17,6 +17,7 @@
     "@radix-ui/react-dialog": "^1.0.3",
     "@radix-ui/react-scroll-area": "^1.0.3",
     "@supabase/supabase-js": "^2.10.0",
+    "@tanstack/react-query": "^4.28.0",
     "@vercel/analytics": "^0.1.11",
     "allotment": "^1.18.1",
     "axios": "^1.3.4",
@@ -33,7 +34,6 @@
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "18.2.0",
     "react-hot-toast": "^2.4.0",
-    "swr": "^2.1.0",
     "zod": "^3.21.4",
     "zustand": "^4.3.6"
   },
