+++ src/components/specific/PageList/PageList.tsx
@@ -5,12 +5,10 @@ import { UsersThree } from "phosphor-react";
 import type { Page } from "@prisma/client";
 import { useSession } from "next-auth/react";
 
-type PageTypes = "characters" | "places" | "items" | "events";
+//* Local imports
+import { useGetPagesByType } from "../../../hooks/specific/useGetPagesByType"
+import type { PageTypes } from "../../../types/page";
 
-type ApiPageListing = {
-  listing: PageTypes;
-  pages: Page[];
-}
 type PageListProps = {
   content: PageTypes;
   worldId: string;
@@ -18,16 +16,8 @@ type PageListProps = {
 
 export default function PageList(props: PageListProps) {
   const title = props.content.charAt(0).toUpperCase() + props.content.slice(1);
-  const [data, setData] = useState<ApiPageListing | null>(null);
 
-  const { data: session } = useSession();
-
-  useEffect(() => {
-    if (session?.user?.id) {
-      getPagesByType(props.worldId, props.content)
-        .then(res => setData(res))
-    }
-  }, []);
+  const pages = useGetPagesByType(props.worldId, props.content);
 
   return (
     <div className="w-full h-full flex flex-col justify-start items-center border-2 border-tertiary-700 rounded-lg overflow-hidden">
@@ -37,7 +27,7 @@ export default function PageList(props: PageListProps) {
       </div>
       <div className="w-full h-full max-h-full min-h-full overflow-x-hidden overflow-y-scroll">
         {
-          data?.pages?.map(page => (
+          pages?.data?.data.pages.map(page => (
             <Link key={page.id} href={`/page/${page.id}`}>
               <button className="w-full px-1 py-2 bg-none hover:bg-tertiary-700 flex flex-row justify-start items-center">
                 <span className="text-tertiary-200 font-bold">{page.name}</span>
@@ -61,24 +51,4 @@ function renderIcon(type: PageTypes) {
     case "events":
       return <UsersThree className="w-8 h-8 text-tertiary-100" size={32} />
   }
-}
-
-async function getPagesByType(worldId: string, pageType: PageTypes) {
-  const body = {
-    worldId,
-    listing: pageType,
-    action: "ListPages"
-  }
-
-  const res = await fetch("/api/pages", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(body),
-  });
-
-  const data = await res.json();
-  console.log(data);
-  return data as ApiPageListing;
 }
