+++ src/types/page.d.ts
@@ -1,0 +1,1 @@
+export type PageTypes = "characters" | "places" | "items" | "events";
